% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proportions.R
\name{add_proportions}
\alias{add_proportions}
\title{Add Proportions}
\usage{
add_proportions(expstudy, ..., .base_grp_nms = character(0), .min_ungrpd = 0L)
}
\arguments{
\item{expstudy}{an \code{\link[=expstudy]{expstudy()}}}

\item{...}{other columns other than the \code{expstudy} metric variables (\code{actuals},
\code{expecteds}, and \code{exposures}) to generate proportions of}

\item{.base_grp_nms}{character vector of column names to use as the base of added proportions}

\item{.min_ungrpd}{minimum number of non-grouping columns required before using
\code{.base_grp_nms} as proportion base}
}
\value{
An \code{expstudy} with added proportions.
}
\description{
Add proportions of \code{expstudy} metric variables. Other variables can be
provided via the <\code{\link[rlang:dyn-dots]{dynamic-dots}}> argument if proportions
are wanted as well. Proportions will sum to 100\% per group if \code{.base_grp_nms}
are provided. If \code{.base_grp_nms} are omitted, proportions will sum to 100\%
in total.
}
