% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{add_proportions}
\alias{add_metrics}
\alias{add_credibility}
\alias{compile_results}
\alias{aggregate}
\alias{expstudy}
\alias{format_metrics}
\title{Defunct functions}
\usage{
# Deprecated in 2.0.0 -------------------------------------

add_proportions(expstudy, ..., .base_grp_nms = character(0), .min_ungrpd = 0L)

add_metrics(
  expstudy,
  ...,
  .metrics = c("act2expec", "act2expos", "expec2expos"),
  .metric_nms = list(act2expec = "ACTUAL_TO_EXPECTED", act2expos = "ACTUAL_TO_EXPOSED",
    expec2expos = "EXPECTED_TO_EXPOSED")
)

add_credibility(
  expstudy,
  .cred_k = 0.05,
  .cred_p = 0.95,
  .cred_nms = "CREDIBILITY"
)

compile_results(
  expstudy,
  ...,
  output = c("metrics", "proportions"),
  output_args = list(metrics = NULL, proportions = NULL),
  output_format = c("unformatted", "formatted")
)

aggregate(expstudy, ..., .oth_sum_vars = NULL)

expstudy(data, actuals, expecteds, exposures, variances = NULL, keys = NULL)

format_metrics(expstudy)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions were part of the initial release of \code{expstudy} but are now
defunct. If there's a known replacement, calling the function
will tell you about it.
}
\keyword{internal}
