% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expstudy.R
\name{is_expstudy}
\alias{is_expstudy}
\title{Test if the object is an expstudy}
\usage{
is_expstudy(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tbl_es} class.
}
\description{
This function returns \code{TRUE} for expstudies or subclasses thereof,
and \code{FALSE} for all other objects.
}
