% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_meta}
\alias{update_meta}
\title{Update expstudy meta data}
\usage{
update_meta(new, old)
}
\arguments{
\item{expstudy}{an \code{expstudy()} object}
}
\description{
Passing an \code{expstudy} through a function often causes the attributes to be
dropped. This is a helper function that retains original attributes.
}
\keyword{internal}
