% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expstudy-package.R
\docType{package}
\name{expstudy-package}
\alias{expstudy-package}
\title{expstudy: Tools for Actuarial Experience Studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Experiences studies are an integral component of the actuarial control cycle. Regardless of the decrement or policyholder behaviour of interest, the analyses conducted is often the same. Ultimately, this package aims to reduce time spent writing the same code used for different experience studies, therefore increasing the time for to uncover new insights inherit within the relevant experience.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cb12991/expstudy}
  \item \url{https://cb12991.github.io/expstudy/}
  \item Report bugs at \url{https://github.com/cb12991/expstudy/issues}
}

}
\author{
\strong{Maintainer}: Cody Buehler \email{cb12991@me.com}

}
\keyword{internal}
