% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expstudy-package.R
\docType{package}
\name{expstudy-package}
\alias{expstudy-package}
\title{expstudy: experience study tools for analytics and communication}
\description{
A set of tools to help aid in the formation and analyses of recurring or novel experience studies.
}
\details{
The \code{expstudy} package provides an subclass, \link{tbl_es}, for
\code{\link[dtplyr:lazy_dt]{dtplyr::lazy_dt()}} that stores attributes relating to an experience study.
These attributes along with other functions reduces the time needed to
review an assumption's efficacy via actual-to-expected analysis in addition
to exploring new relevant patterns or correlations.
}
\author{
\strong{Maintainer}: Cody Buehler \email{cb12991@me.com}

}
\keyword{internal}
