% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metrics.R
\name{format_metrics}
\alias{format_metrics}
\title{Format \code{expstudy} metrics}
\usage{
format_metrics(expstudy)
}
\arguments{
\item{expstudy}{an \code{expstudy} object}
}
\value{
the same \code{expstudy} with formatted metrics
}
\description{
Easily format metrics and metric variables for better readability.
}
