% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{collect}
\alias{compute}
\alias{count}
\alias{distinct}
\alias{filter}
\alias{glimpse}
\alias{group_by}
\alias{group_map}
\alias{group_modify}
\alias{intersect}
\alias{left_join}
\alias{mutate}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{select}
\alias{setdiff}
\alias{slice}
\alias{slice_head}
\alias{slice_max}
\alias{slice_min}
\alias{slice_tail}
\alias{summarise}
\alias{transmute}
\alias{ungroup}
\alias{union}
\alias{union_all}
\alias{complete}
\alias{drop_na}
\alias{expand}
\alias{fill}
\alias{nest}
\alias{pivot_longer}
\alias{pivot_wider}
\alias{replace_na}
\alias{separate}
\alias{head}
\alias{tail}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr:compute]{collect}}, \code{\link[dplyr]{compute}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_map}}, \code{\link[dplyr:group_map]{group_modify}}, \code{\link[dplyr:reexports]{intersect}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:reexports]{setdiff}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:mutate]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}, \code{\link[dplyr:reexports]{union}}, \code{\link[dplyr:setops]{union_all}}}

  \item{pillar}{\code{\link[pillar]{glimpse}}}

  \item{tidyr}{\code{\link[tidyr]{complete}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{expand}}, \code{\link[tidyr]{fill}}, \code{\link[tidyr]{nest}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{replace_na}}, \code{\link[tidyr]{separate}}}

  \item{utils}{\code{\link[utils]{head}}, \code{\link[utils:head]{tail}}}
}}

